refactor(plan_SERVERSPLIT slice 4): the dynamic-hooks bodies leave server.py
The ten `dynamic-hooks` tool bodies move VERBATIM out of `embedded/server.py`
into `embedded/tool_sections/dynamic_hooks.py`. No behaviour change, no rename,
no tool added or removed: the served surface is byte-identical to
`tests/goldens/tool_surface.json` (the HARD golden), and all three loaded
identities of `server.py` still build a full 94-tool app.

This is the section the plan orders fourth for one reason: it is the ONLY one
with synchronous tool bodies. `tool_registry._surrogate_safe_returns` branches on
`inspect.iscoroutinefunction` and wraps a sync tool in a sync wrapper; after the
move the binding loop applies that wrapper from `server.py` to a function whose
`__globals__` is a section module. `tests/test_tool_dispatch.py` pins that the
registered `get_hook_documentation` is still not a coroutine function, and the
94-minus-5 async split is unmoved. Beyond `rt.dynamic_hook_ai` the section has no
shared dependencies at all — no tab, no browser manager, no CDP timeout — so the
sync branch is proven in isolation before the two heavier sections follow.

The only edits to a moved body are the dropped registration decorator (contract
rule 2) and the rewrite of the singleton read to `rt.<name>`, resolved at CALL
time against `tool_runtime`, the one patchable home (contract rule 3).

server.py: 3014 -> 2839 LOC; the GRANDFATHER cap ratchets DOWN to that measured
post-`ruff format` actual, and `tests/source_scan.py`'s floor ratchets UP 5 -> 6.
The one-line `dynamic_hook_ai` alias import server.py was the last consumer of is
pruned with the bodies.

Verified: unit lane 2333 passed / 1 skipped (base 2334 — the difference is one
parametrize case of the migration alias-identity pin, which enumerates
`tool_runtime.__all__` ∩ `dir(server)` and shrinks by one with every pruned
alias); wire surface IDENTICAL; three-identity load SECTION_TOOLS 94 / per-app
[94, 94, 94]; `--list-sections` subprocess "Total: 94 tools"; ruff format+check,
budgets, suppression owners, vulture, ty (77 warnings, unchanged) and
`gen_release_contract --check` all clean; real-Chrome `test_e2e_functions_hooks.py`
7 passed and real-stdio `test_e2e_transport.py` 1 passed.
