+++ pyproject.toml
@@ -256,6 +256,15 @@ ban-relative-imports = "all"
 # is the exact subset of server.py's file-wide list the moved lines fire —
 # notably A002 for export_debug_logs' `format` parameter, which is part of the
 # WIRE surface and so cannot be renamed by a pure move. Carried verbatim.)
+"src/stealth_chrome_devtools_mcp/embedded/tool_sections/dynamic_hooks.py" = [
+    "E501", "PLR0913", "TID251",
+]  # plan_SERVERSPLIT slice 4 (the dynamic-hooks bodies moved VERBATIM out of
+# server.py — the only section with SYNCHRONOUS tool bodies. Each code is the
+# exact subset of server.py's file-wide list the moved lines fire: TID251 the
+# `typing.Any` annotations, PLR0913 create_simple_dynamic_hook's six template
+# fields, E501 three docstring lines in create_dynamic_hook's Args block. The
+# docstrings are byte-identical to server.py's because FastMCP serves them and
+# tests/goldens/tool_surface.json is a HARD golden. Carried verbatim.)
 "src/stealth_chrome_devtools_mcp/embedded/clone_storage.py" = [
     "ANN001", "ANN202", "ANN401", "ARG001", "ARG005", "ASYNC240",
     "BLE001", "DTZ005", "DTZ006", "E501", "PIE810", "PLR2004", "PTH100",
